requirements.txt
gunicorn
bottle
bottle-sqlalchemy
mako
redis
bottle-redis
